coffee/app/requirements.txt
aiogram==3.25.0
openai==1.97.0
httpx==0.28.1
requests==2.32.3
sqlalchemy==2.0.40
aiosqlite==0.22.0
fastapi==0.116.1
uvicorn==0.34.0
sqladmin==0.20.1
python-dotenv==1.1.0
python-multipart==0.0.20
pytest==8.3.5
pytest-asyncio==0.25.3
hypothesis==6.125.2
aioresponses==0.7.8
pytest-mock==3.14.0
Pillow==12.1.1
bcrypt==4.3.0
pyjwt==2.10.1
python-dateutil==2.9.0.post0
